Make coqsplit build with -safe-string
OCaml 4.06.0 (November 2017) made -safe-string the default. Without this
change, building coqsplit with a newer ocamlc fails with

Error: This expression has type bytes but an expression was expected of type
         string

regarding the second argument to apply_spec in main's definition of s1.

The Bytes module along with the -safe-string and -unsafe-string options
were introduced with OCaml 4.02.0 (August 2014).